Reorganize port cleanup
